% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_coloc.R
\name{addAssociationPosterior}
\alias{addAssociationPosterior}
\title{Calculate posterior probabilities from association summary stats}
\usage{
addAssociationPosterior(dataset, n)
}
\arguments{
\item{dataset}{Association summary statistics
(required columns: snp_id, p_nominal, effect_size, MAF).}

\item{n}{Sample size of the dataset}
}
\value{
Original data frame with one additional column containing posterior probabilities.
}
\description{
Calculate posterior probabilities from association summary stats
}
