% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_postProcess.R
\name{addExpectedPvalue}
\alias{addExpectedPvalue}
\title{Add expected p-value for Q-Q plots}
\usage{
addExpectedPvalue(pvalue_df)
}
\arguments{
\item{pvalue_df}{}
}
\value{
pvalue_df with p_expected column
}
\description{
Add expected p-value for Q-Q plots
}
