% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_classify_peaks.R
\name{addOverlappingPeaks}
\alias{addOverlappingPeaks}
\title{Add overlapping ATAC peaks to a credible set of QTL variants}
\usage{
addOverlappingPeaks(qtl_credible_set, atac_peak_metadata, extend = 0)
}
\arguments{
\item{qtl_credible_set}{Data frame with credible set of QTL variants 
(columns: snp_id, chr, pos)}

\item{atac_peak_metadata}{Data frame of start end coordinates of ATAC peaks 
(columns: gene_id, chr, start, end, strand)}

\item{extend}{Number of bp added to peak start and end coordinates to increase the chance of finding an overlap.}
}
\value{
Modifed qtl_credible_set data frame with added overlap_peak_id column.
}
\description{
Add overlapping ATAC peaks to a credible set of QTL variants
}
