% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_postProcess.R
\name{addR2FromLead}
\alias{addR2FromLead}
\title{Calculate R2 between the lead SNP and all other SNPs in the data frame and adds corresponding column.}
\usage{
addR2FromLead(gene_df, genotypes)
}
\arguments{
\item{gene_df}{Data frame of QTL snps (required column: gene_id)}

\item{genotypes}{Genotype matrix}
}
\value{
gene_df data frame with R2 column added.
}
\description{
Assumes that lead SNP is in the first row of the gene_df data frame 
(data frame is sorted by p-value).
}
