% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_sequences.R
\docType{data}
\name{alaska_232_2015_nts}
\alias{alaska_232_2015_nts}
\title{1650 nt HA sequence for Alaska/232/2015 (useful for aligning)}
\format{
An object of class \code{character} of length 1.
}
\usage{
alaska_232_2015_nts
}
\description{
1650 nt HA sequence for Alaska/232/2015 (useful for aligning)
}
\keyword{datasets}
