% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_finemapping.R
\name{annotateCredibleSet}
\alias{annotateCredibleSet}
\title{Add ATAC peak annotations to the eQTL credible set}
\usage{
annotateCredibleSet(eqtl_credible_set, atac_peak_metadata, atac_tabix_path)
}
\arguments{
\item{eqtl_credible_set}{}

\item{atac_peak_metadata}{}

\item{atac_tabix_path}{}
}
\value{
Credible set with ATAC peak annotations
}
\description{
Add ATAC peak annotations to the eQTL credible set
}
