% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_substitutions.R
\name{apply_substitutions}
\alias{apply_substitutions}
\title{Apply substitutions to a sequence}
\usage{
apply_substitutions(sequence, substitutions, type = "aa")
}
\arguments{
\item{sequence}{Character string of amino acids or nucleotides}

\item{substitutions}{Character vector of substitutions in format "K2N" (change K at position 2 to N)}

\item{type}{Sequence type: "aa" for amino acids or "nt" for nucleotides}
}
\value{
Modified sequence with substitutions applied
}
\description{
Apply substitutions to a sequence
}
