% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_ASE.R
\name{aseDataAddGenotypes}
\alias{aseDataAddGenotypes}
\title{Title Add genotype data to tidy ASE data frame.}
\usage{
aseDataAddGenotypes(ase_data, genotypes)
}
\arguments{
\item{ase_data}{Tidy ASE data frame with at least the followign columns: feature_snp_id, lead_snp_id, genptype_id.}

\item{genotypes}{Standard genotype matrix (variants in rows, individuals in columns)}
}
\value{
ASE data frame that contains genotypes for each individual for each feature snp and lead snp.
}
\description{
Title Add genotype data to tidy ASE data frame.
}

