% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_gwas.R
\name{calculateEnrichment}
\alias{calculateEnrichment}
\title{Calculate the enrichment of GWAS traits among eQTLs}
\usage{
calculateEnrichment(gwas_catalog, gwas_snp_pvalues, eqtl_lead_snps,
  gene_id_name_map, n_genes = 14163, p_diff_thresh = 3,
  min_genes_per_trait = 20)
}
\arguments{
\item{gwas_catalog}{EBI GWAS catalog imported using the importGwasCatalog function.}

\item{gwas_snp_pvalues}{FastQTL eQTL pvalues for each SNP in the GWAS catalog, from importGwasPvalue.}

\item{eqtl_lead_snps}{Lead eQTL SNPs for each significant gene.}

\item{gene_id_name_map}{- data frame with two columns: gene_id, gene_name.}

\item{n_genes}{Total numner of genes used in the eQTL analysis.}

\item{p_diff_thresh}{Log10 p-value difference between lead eQTL SNP and lead GWAS SNP.}

\item{min_genes_per_trait}{Minimal number of unique genes per trait to be used for enrichment analysis.}
}
\value{
list(enrichment = traits enriched among eQTLs, overlap = table of overlapping genes and traits).
}
\description{
Based on Supplementary Figure S16 in (Kumasaka et al 2015).
}
\author{
Kaur Alasoo
}
