% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_processData.R
\name{calculatePairwisePi1}
\alias{calculatePairwisePi1}
\title{Calculate all pairwise Pi1 statistics for a list p-value tables.}
\usage{
calculatePairwisePi1(qtl_list, qvalue_thresh = 0.1, tidy = FALSE)
}
\arguments{
\item{qtl_list}{List of p-value tables}

\item{qvalue_thresh}{qvalue threshold for table1.}
}
\value{
None
}
\description{
Currently expects the following columns: gene_id, qvalue and p_beta.
}
\author{
Kaur Alasoo
}

