% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_processPeaks.R
\name{calculatePeakDistance}
\alias{calculatePeakDistance}
\title{Caclulate distance between the centres of peak pairs}
\usage{
calculatePeakDistance(peak_df, peak_metadata)
}
\arguments{
\item{peak_df}{Data frame with two columns containing peak_ids (required columns: master_id, dependent_id)}

\item{peak_metadata}{Matrix with peak metadata (required_columns: gene_id, chr, start, end)}
}
\value{
peak_df with dependent_centre, master_centre and distance column added.
}
\description{
Caclulate distance between the centres of peak pairs
}

