% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_processData.R
\name{calculatePi1}
\alias{calculatePi1}
\title{Calculate Pi1 replicability statistic between two sets of pvalues}
\usage{
calculatePi1(table1, table2, qvalue_thresh = 0.1)
}
\arguments{
\item{table1}{First table maximum p-values per feature.}

\item{table2}{Second table of maximum p-values per feature.}

\item{qvalue_thresh}{qvalue threshold for table1.}
}
\value{
None
}
\description{
Currently expects the following columns: gene_id, qvalue and p_beta.
}
\author{
Kaur Alasoo
}

