% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_munging.R
\name{calculateTPM}
\alias{calculateTPM}
\title{Normalise counts matrix using transcripts per million (TPM) normalisation}
\usage{
calculateTPM(counts_matrix, lengths, selected_genes = NULL,
  fragment_length = 250)
}
\arguments{
\item{counts_matrix}{Matrix of read counts (genes in rows, samples in columns)}

\item{lengths}{data frame with gene lengths (required columns: gene_id, length)}

\item{selected_genes}{Subset of genes used to to calculate library size (default: NULL = use all genes)}

\item{fragment_length}{Mean fragment length}
}
\value{
Read count matrix where entries have been TPM normalized.
}
\description{
Normalise counts matrix using transcripts per million (TPM) normalisation
}

