% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_sequences.R
\docType{data}
\name{california_04_2009_h1_aas}
\alias{california_04_2009_h1_aas}
\title{aa HA sequence for Alaska/232/2015 (useful for aligning)}
\format{
An object of class \code{character} of length 1.
}
\usage{
california_04_2009_h1_aas
}
\description{
aa HA sequence for Alaska/232/2015 (useful for aligning)
}
\keyword{datasets}
