% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment.R
\name{censorLog2OR}
\alias{censorLog2OR}
\title{Censor log2 odds ratios and their confidence intervals for easier plotting}
\usage{
censorLog2OR(or_df, min_OR_log2 = -3, min_CI_lower_log2 = -4)
}
\arguments{
\item{or_df}{data frame of ORs}

\item{min_OR_log2}{}

\item{min_CI_lower_log2}{}
}
\value{
Modified or_df with censored values
}
\description{
Censor log2 odds ratios and their confidence intervals for easier plotting
}

