% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_coloc.R
\name{colocMolecularQTLs}
\alias{colocMolecularQTLs}
\title{Perform colocalisation between an single QTL and GWAS summary stats from the same region.}
\usage{
colocMolecularQTLs(qtl_df, qtl_summary_path, gwas_summary_path,
  gwas_variant_info, qtl_variant_info, N_qtl = 84, cis_dist = 1e+05,
  QTLTools = TRUE)
}
\arguments{
\item{qtl_df}{A data frame with a single row and two columns (phenotype_id, snp_id) corresponding the a single QTL.}

\item{qtl_summary_path}{Path to the tabix indexed QTL summary statistics file generated by QTLtools}

\item{gwas_summary_path}{Path to the tabix indexed GWAS summary statistics file.}

\item{gwas_variant_info}{Variant information for the GWAS data. Imported using importVariantInformation() function.}

\item{qtl_variant_info}{Variant information for the GWAS data. Imported using importVariantInformation() function. 
The gwas_variant_info and qtl_variant_info data frames are used to link GWAS variants to QTL variants. As a result, they
must have exactly the same values in the snp_id column to allow matching. They can, however, have different values in 
the chr pos columns. This is useful when the the GWAS and QTL mapping have been performed using different versions of the 
reference genome (i.e. GRCh37 vs GRCh38). Consequently, colocalisation can be performed even without 
lifting over the GWAS summary statstitics, it's sufficient to have a different version of the variant infromation file.}

\item{N_qtl}{Sample size for QTL mapping. Used by coloc to estimate the the standard errors from p-values and effect sizes.}

\item{cis_dist}{With of the genomic region around the lead QTL variant that is used for colocalisation; width = 2*cis_dist.}

\item{QTLTools}{Set to TRUE if QTL mapping was performed using QTLtools.}
}
\value{
List of colocalisation results or NULL values if there was an error.
}
\description{
Perform colocalisation between an single QTL and GWAS summary stats from the same region.
}
