% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_coloc.R
\name{colocMolecularQTLsByRow}
\alias{colocMolecularQTLsByRow}
\title{Applies the colocMolecularQTLs function to each row of the qtl_df data frame.}
\usage{
colocMolecularQTLsByRow(qtl_df, ...)
}
\arguments{
\item{qtl_df}{Data frame of QTLs}

\item{...}{Additional parameters passed on to colocMolecularQTLs function}
}
\description{
See documentation for colocMolecularQTLs for more details
}
