% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_coloc.R
\name{colocQtlGWAS}
\alias{colocQtlGWAS}
\title{Test colocalisation between molecular QTL and GWAS summary stats}
\usage{
colocQtlGWAS(qtl, gwas, N_qtl)
}
\arguments{
\item{qtl}{QTL summary stats (p_nominal, MAF, beta, snp_id)}

\item{gwas}{GWAS summary stats(beta, se, MAF, log_OR)}

\item{N_qtl}{Sample size of the QTL mapping study}
}
\value{
coloc.abf result object
}
\description{
Test colocalisation between molecular QTL and GWAS summary stats
}
