% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{conditionPalette}
\alias{conditionPalette}
\title{Colour plaette of four colours to represent experimental conditions}
\usage{
conditionPalette()
}
\description{
Colour plaette of four colours to represent experimental conditions
}
