% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_coloc.R
\name{constructCredibleSet}
\alias{constructCredibleSet}
\title{Construct the credible set of associated SNPs for each gene.}
\usage{
constructCredibleSet(dataset, threshold = 0.95)
}
\arguments{
\item{dataset}{Output from addAssociationPosterior}

\item{threshold}{Cumulative posterior porbability of most associated SNPs.}
}
\value{
Dataset that only contains the credible set of associated SNPs.
}
\description{
Construct the credible set of associated SNPs for each gene.
}

