% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_fastqtl.R
\name{constructFastQTLGenePos}
\alias{constructFastQTLGenePos}
\title{Construct gene position data.frame suitable for FastQTL}
\usage{
constructFastQTLGenePos(gene_metadata)
}
\arguments{
\item{gene_metadata}{Gene metadata data frame from expression list (columns: chr, start, end, gene_id)}
}
\value{
data_frame suitable for FastQTL
}
\description{
Construct gene position data.frame suitable for FastQTL
}
\author{
Kaur Alasoo
}
