% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_list_functions.R
\name{constructGeneData}
\alias{constructGeneData}
\title{For a given gene_id, extract its expression from expression_matrix and join with metadata.}
\usage{
constructGeneData(gene_id, expression_matrix, sample_metadata)
}
\arguments{
\item{gene_id}{Gene id, corresponds to a row name of expression matrix.}

\item{expression_matrix}{Matrix of gene exrpression; rows - gene ids, cols - sample_ids.}

\item{sample_metadata}{Data frame with metadata for each sample.}
}
\value{
data frame that has gene expression in value column and metadata in other columns.
}
\description{
For a given gene_id, extract its expression from expression_matrix and join with metadata.
}
\author{
Kaur Alasoo
}
