% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{constructGeneRanges}
\alias{constructGeneRanges}
\title{Construct GRanges object for tabixFetchGenes to fetch cis regions around each gene.}
\usage{
constructGeneRanges(selected_genes, gene_metadata, cis_window)
}
\arguments{
\item{selected_genes}{data frame containing at least gene_id column.}

\item{gene_metadata}{data frame with gene metadata (required columns: gene_id, chr, start, end)}

\item{cis_window}{Size fo the cis-window around the gene}
}
\value{
GRanges object with cooridnates of cis regions around genes.
}
\description{
Construct GRanges object for tabixFetchGenes to fetch cis regions around each gene.
}

