% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{constructGenotypeText}
\alias{constructGenotypeText}
\title{Construct matching between minor allele count and genotype text}
\usage{
constructGenotypeText(snp_id, variant_information)
}
\description{
Construct matching between minor allele count and genotype text
}
