% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_matrixEQTL.R
\name{constructMatrixEQTLCovariates}
\alias{constructMatrixEQTLCovariates}
\title{Convert sample_metadata into a covariates matrix suitable for matrixEQTL}
\usage{
constructMatrixEQTLCovariates(sample_metadata, covariate_names,
  id_column = "sample_id")
}
\arguments{
\item{sample_metadata}{Sample metadata matrix}

\item{covariate_names}{Vector of columns names that will be used as covariates.}

\item{id_column}{Name of the column that will be used as column names for the covariates matrix.
Default sample id.}
}
\value{
Covariates matrix suitable for matrixEQTL.
}
\description{
Convert sample_metadata into a covariates matrix suitable for matrixEQTL
}

