% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_matrixEQTL.R
\name{constructMatrixEQTLGenePos}
\alias{constructMatrixEQTLGenePos}
\title{Convert gene_metadata table into a gene positions data frame suitable for MatrixEQTL.}
\usage{
constructMatrixEQTLGenePos(gene_metadata)
}
\arguments{
\item{gene_metadata}{Data frame with gene metadata. Required columns: gene_id, chr, start, end)}
}
\value{
Gene postion data frame for matrixEQTL.
}
\description{
Convert gene_metadata table into a gene positions data frame suitable for MatrixEQTL.
}
