% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{constructQTLtoolsGenePos}
\alias{constructQTLtoolsGenePos}
\title{Construct gene position data.frame suitable for QTLtools}
\usage{
constructQTLtoolsGenePos(gene_metadata)
}
\arguments{
\item{gene_metadata}{Gene metadata data frame from expression list (columns: chr, start, end, gene_id, transcript_id, strand)}
}
\value{
data_frame suitable for FastQTL
}
\description{
Construct gene position data.frame suitable for QTLtools
}
\author{
Kaur Alasoo
}
