% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_coloc.R
\name{constructVariantRanges}
\alias{constructVariantRanges}
\title{Construct a GRanges obejct corresponding to a cis region around one variant.}
\usage{
constructVariantRanges(variant_df, variant_information, cis_dist)
}
\arguments{
\item{variant_df}{data frame with at least snp_id column}

\item{variant_information}{data.frame from importVariantInformation() function}

\item{cis_dist}{Number of basepairs upstream and downstream of the variant.}
}
\value{
GRanges
}
\description{
Construct a GRanges obejct corresponding to a cis region around one variant.
}
