% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{countSnpsOverlapingExons}
\alias{countSnpsOverlapingExons}
\title{Count the number of feature SNPs and cis SNPs overlapping exons of all genes}
\usage{
countSnpsOverlapingExons(gene_metadata, snp_coords, cis_window = 5e+05)
}
\arguments{
\item{gene_metadata}{Data frame with gene metadata (required columns: gene_id, chr, strand, exon_starts, exon_ends)
-exon_stars: comma-separated list of exon start coordinates
-exon_ends: comma-separated list of exon end coordinates.}

\item{snp_coords}{Data frame with SNP coordinates from a VCF file (required columns: chr, pos, snp_id)}

\item{cis_window}{Size of the cis window from both sides of the gene.}
}
\value{
Data frame with exon coordinates, cis region coordiantes as well as the number of cis and feature snps.
}
\description{
Count the number of feature SNPs and cis SNPs overlapping exons of all genes
}

