% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{countSnpsOverlapingPeaks}
\alias{countSnpsOverlapingPeaks}
\title{Count the number of feature SNPs and cis SNPs overlapping a set of peak calls}
\usage{
countSnpsOverlapingPeaks(peak_metadata, snp_coords, cis_window = 5e+05)
}
\arguments{
\item{peak_metadata}{Data frame with peak metadata (required columns: gene_id, chr, start, end)}

\item{snp_coords}{Data frame with SNP coordinates from a VCF file (required columns: chr, pos, snp_id)}

\item{cis_window}{Size of the cis window from both sides of the peak.}
}
\value{
Data frame with peak coordinates, cis region coordiantes as well as number of cis and feature snps.
}
\description{
Count the number of feature SNPs and cis SNPs overlapping a set of peak calls
}

