% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_classify_peaks.R
\name{credibleSetPeakOverlaps}
\alias{credibleSetPeakOverlaps}
\title{Identify all peak-peak pairs linked by credible sets}
\usage{
credibleSetPeakOverlaps(credible_sets_df)
}
\description{
Identify all peak-peak pairs linked by credible sets
}
