% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_processGeneAnnotations.R
\name{dataFrameToGRanges}
\alias{dataFrameToGRanges}
\title{Convert a data frame into a GRanges object}
\usage{
dataFrameToGRanges(df)
}
\arguments{
\item{df}{Input data frame (required columns: seqnames, start, end, strand)}
}
\value{
GRanges object construct from df.
}
\description{
Seqnames, strand, start and end columns are used as corresponding elements 
in the GRanges object. Remaining columns are added into the elementMetadata data frame.
}

