% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_eigenMT.R
\name{eigenMTExportGenotypes}
\alias{eigenMTExportGenotypes}
\title{Export genotype data in format suitable for eigenMT}
\usage{
eigenMTExportGenotypes(vcf_file, eigenMT_dir, prefix)
}
\arguments{
\item{vcf_file}{Genotype data imported with gdsToMatrix}

\item{eigenMT_dir}{Directory of the eigenMT input files}

\item{prefix}{Prefix of the output files,}
}
\description{
Creates two text files in the eigenMT_dir directory: [prefix].snp_positions.txt and 
 [prefix].genotypes.txt.
}

