% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_eigenMT.R
\name{eigenMTExportGenotypesByChr}
\alias{eigenMTExportGenotypesByChr}
\title{Convert chromosome-wise .gds files into genotype matrices suitable for eigenMT}
\usage{
eigenMTExportGenotypesByChr(chromosomes, gds_dir, eigenMT_dir,
  gds_prefix = "chr_")
}
\arguments{
\item{chromosomes}{Vector of chromosome names.}

\item{gds_dir}{Directory containing the chromosome .gds files.}

\item{eigenMT_dir}{Directory for the output files.}

\item{gds_prefix}{Prefix of the .gds files}
}
\description{
Convert chromosome-wise .gds files into genotype matrices suitable for eigenMT
}
