% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_eigenMT.R
\name{eigenMTImportResults}
\alias{eigenMTImportResults}
\title{Import eigenMT output file into data frame}
\usage{
eigenMTImportResults(file_path)
}
\arguments{
\item{Path}{to the eigenMT output file.}
}
\value{
Data frame of eigenMT corrected p-values
}
\description{
Also performs FDR correction
}
