% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VarCompFunctions.R
\name{estimateVarianceExplained}
\alias{estimateVarianceExplained}
\title{Estimate propotion of variance explained by each compoent in a model.}
\usage{
estimateVarianceExplained(model_data, model_function)
}
\arguments{
\item{model_data}{data frame with all of the neccessary columns for the model.}

\item{model_function}{Function that when applied to a model_data data frame returns an
lme4 object.}
}
\value{
data frame containing the proportion of variance explained by each component.
}
\description{
#Apply an lme4 model (model_function) to a data frame and
report the proportion of variance explained by each component.
}
\author{
Kaur Alasoo
}

