% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example}
\alias{example}
\title{Toy Example Expression Value Data}
\format{
an object of class \code{"matrix"} and \code{"array"}
}
\usage{
data(example)
}
\description{
This toy example is for demonstrating the format in which expression values
should be presented for using functions in this package. It is generated using
the following code:
}
\details{
\code{example <- matrix(round(rchisq(1000, 10^3)), 100, 10)}
\code{rownames(example) <- paste('gene', seq_len(100), sep = '_')}
\code{colnames(example) <- paste('sample', seq_len(10), sep = '_')}
}
\examples{
data(example)
dim(example)
rownames(example)[1:10]
colnames(example)[1:10]
}
\keyword{dataset}
\keyword{example}
