% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{exportDataForRasqual}
\alias{exportDataForRasqual}
\title{Write multiple files for RASQUAL onto disk.}
\usage{
exportDataForRasqual(condition_list, rasqual_input_folder,
  max_batch_size = 50)
}
\arguments{
\item{condition_list}{Named list of expression lists, each expression list needs to
contain at least the following elements: 'counts' matrix, 'sample_metadata' df,}

\item{rasqual_input_folder}{Path to the RASQUAL input folder.}

\item{max_batch_size}{Maximal number of feaures to be included in a single batch.}
}
\description{
Write multiple files for RASQUAL onto disk.
}

