% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_list_functions.R
\name{extractConditionFromExpressionList}
\alias{extractConditionFromExpressionList}
\title{Extract samples from expression list based on condition_name column in metadata.}
\usage{
extractConditionFromExpressionList(cond_name, expression_list)
}
\arguments{
\item{cond_name}{Name of the condition, has to match at least one entry in the 
condition_name columns of the sample_metadata data frame.}

\item{expression_list}{Expression list to be filtered.}
}
\value{
Filtered expression list
}
\description{
Extract samples from expression list based on condition_name column in metadata.
}
