% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_list_functions.R
\name{extractConditionFromSummarizedExperiment}
\alias{extractConditionFromSummarizedExperiment}
\title{Extract samples from SummarizedExperiment based on condition_name column in metadata.}
\usage{
extractConditionFromSummarizedExperiment(cond_name, summarizedExperiment)
}
\arguments{
\item{cond_name}{Name of the condition, has to match at least one entry in the 
condition_name columns of the colData data frame.}

\item{expression_list}{Expression list to be filtered.}
}
\value{
Filtered SummarizedExperiment
}
\description{
Extract samples from SummarizedExperiment based on condition_name column in metadata.
}

