% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_list_functions.R
\name{extractGenesFromExpressionList}
\alias{extractGenesFromExpressionList}
\title{Extract a subset of genes from expression list}
\usage{
extractGenesFromExpressionList(expression_list, gene_ids)
}
\arguments{
\item{expression_list}{Input expression list object.}

\item{gene_ids}{Vector of gene ids to be retained.}
}
\value{
Filtered expression list that only contains genes from gene_ids.
}
\description{
Extract a subset of genes from expression list
}
