% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{fast_fasta}
\alias{fast_fasta}
\title{Read a fasta file quite quickly using \code{readr::read_lines()}}
\usage{
fast_fasta(path)
}
\arguments{
\item{path}{path of .fasta to read}
}
\value{
Named character vector of sequences
}
\description{
Read a fasta file quite quickly using \code{readr::read_lines()}
}
