% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{fast_fasta.write}
\alias{fast_fasta.write}
\title{Write a fasta file quite quickly using \code{readr::write_lines()}}
\usage{
fast_fasta.write(seqs, names = NULL, path)
}
\arguments{
\item{seqs}{a named or unnamed (in which case \code{names} must be provided) character vector of sequences}

\item{names}{(optional) names of sequences, required if \code{sequences} is unnamed.}

\item{path}{path to write .fasta}
}
\description{
Write a fasta file quite quickly using \code{readr::write_lines()}
}
