% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_fastqtl.R
\name{fastqtlTabixFetchGenes}
\alias{fastqtlTabixFetchGenes}
\title{Fetch particular genes from tabix indexed FastQTL output file.}
\usage{
fastqtlTabixFetchGenes(gene_ranges, tabix_file)
}
\arguments{
\item{gene_ranges}{GRanges object with coordinates of the cis regions around genes.}

\item{tabix_file}{Tabix-indexed fastqtl output file.}
}
\value{
List of data frames containing Rasqual results for each gene.
}
\description{
Fetch particular genes from tabix indexed FastQTL output file.
}

