% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_fastqtl.R
\name{fastqtlTabixFetchSNPs}
\alias{fastqtlTabixFetchSNPs}
\title{Fetch particular SNPs from tabix indexed Fastqtl output file.}
\usage{
fastqtlTabixFetchSNPs(snp_ranges, tabix_file)
}
\arguments{
\item{snp_ranges}{GRanges object with SNP coordinates.}

\item{tabix_file}{Tabix-indexed FastQTL output file.}
}
\value{
Data frame that contains all tested rasqual p-values fir each SNP.
}
\description{
Fetch particular SNPs from tabix indexed Fastqtl output file.
}
