% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{fetchMultipleGenes}
\alias{fetchMultipleGenes}
\title{Fetch multiple genes from SQLite database
NOTE: This implementation is extremely inefficient}
\usage{
fetchMultipleGenes(gene_ids, db_table)
}
\description{
Fetch multiple genes from SQLite database
NOTE: This implementation is extremely inefficient
}

