% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{fetchMultipleGeneSNPPairs}
\alias{fetchMultipleGeneSNPPairs}
\title{Fetch multiple genes from SQLite database
NOTE: This implementation is extremely inefficient}
\usage{
fetchMultipleGeneSNPPairs(gene_snp_pairs, db_table)
}
\description{
Fetch multiple genes from SQLite database
NOTE: This implementation is extremely inefficient
}

