% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{fetchMultipleSNPs}
\alias{fetchMultipleSNPs}
\title{Fetch multiple SNPs from SQLite database
NOTE: This implementation is extremely inefficient}
\usage{
fetchMultipleSNPs(snp_df, db_table)
}
\description{
Fetch multiple SNPs from SQLite database
NOTE: This implementation is extremely inefficient
}

