% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{fetchSQLite}
\alias{fetchSQLite}
\title{Extract rasqual results for particular gene_id or snp_id from the SQLite database}
\usage{
fetchSQLite(db_table, selected_gene_id = NULL, selected_snp_id = NULL)
}
\arguments{
\item{db_table}{Connection to SQLite database table.
(required columns: gene_id, snp_id, p_nominal, allele_freq).}

\item{selected_gene_id}{Id of the selected gene.}

\item{selected_snp_id}{Id of the selected SNP.}
}
\value{
data.frame with rasqual results for a particular gene or SNP.
}
\description{
Either selected_gene_id or selected_snp_id must be specified.
}

