% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_postProcess.R
\name{filterHitsR2}
\alias{filterHitsR2}
\title{Filter SNPs based on R2}
\usage{
filterHitsR2(feature_snp_pairs, genotypes, R2_thresh = 0.8)
}
\arguments{
\item{feature_snp_pairs}{Data frame with two columns: gene_id, snp_id.}

\item{genotypes}{Genotype matrix (SNPs in rows, samples in columns)}

\item{R2_thresh}{R2 threshold, default 0.8.}
}
\value{
Filtered feature_snp_pairs that does not cotain highly correlated SNPs for single feature.
}
\description{
Identify genes that have two different lead SNPs and remove one of them if 
the R2 between the two SNPs is greater than R2_thresh.
}
