% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_test_interactions.R
\name{filterInteractionResults}
\alias{filterInteractionResults}
\title{Filter QTL interaction test results based on RASQUAL effect size}
\usage{
filterInteractionResults(conditions, interaction_result, pvalue_list,
  fdr_thresh = 0.1)
}
\arguments{
\item{conditions}{Character vector of condition names.}

\item{interaction_result}{Output from the matrixeqtlTestInteraction function.}

\item{pvalue_list}{List of RASQUAL gene-SNP results by condition.}

\item{fdr_thresh}{Interaction test FDR threshold.}
}
\value{
Data frame augmented with Rasqual effect sizes and p-values in each condition.
}
\description{
Filter QTL interaction test results based on RASQUAL effect size
}

