% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annot_processPeaks.R
\name{filterOverlaps}
\alias{filterOverlaps}
\title{Filter peak calls based on how many other samples they are found in.}
\usage{
filterOverlaps(peak_list, minOverlapCount = 3)
}
\arguments{
\item{peak_list}{list of GRanges objects corresponding to peak calls.}

\item{minOverlapCount}{Minimal number of samples the peak has to be present to be retained.}
}
\value{
List of GRanges objects corresponding to filtered peak calls from each sample, only peaks
that were present in minOverlapCount samples are retained.
}
\description{
For each set of peak calls, find the ones that are present in at least minOverlapCount samples.
}
\author{
Kaur Alasoo
}

