% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment.R
\name{fimoRelativeEnrichment}
\alias{fimoRelativeEnrichment}
\title{Calculate relative motif enrichment in foreground set of peaks compared to a specifed background.}
\usage{
fimoRelativeEnrichment(foreground, background = NULL, fimo_hits,
  peak_metadata)
}
\arguments{
\item{foreground}{Data frame of foreground peaks, should be a subset of background peaks.}

\item{background}{Data frame of background peaks (deafult: NULL). If NULL then uses all peaks in peak_metadata.}

\item{fimo_hits}{Data frame containg all FIMO motif matches in all peaks.}

\item{peak_metadata}{Data frame with peak start and end coordinates.}
}
\value{
Data frame with relative enrichment of each motif inf the foreground set compared to the background set.
}
\description{
Calculate relative motif enrichment in foreground set of peaks compared to a specifed background.
}
