% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_gwas.R
\name{findGWASOverlaps}
\alias{findGWASOverlaps}
\title{Find GWAS variants that are in high LD with QTL variants}
\usage{
findGWASOverlaps(gene_snp_pairs, gwas_catalog, vcf_file,
  max_distance = 1e+06, min_r2 = 0.6)
}
\arguments{
\item{gene_snp_pairs}{QTL SNP and variant pairs (required columns: gene_id, snp_id)}

\item{gwas_catalog}{GWAS catalog data_frame from the importGwasCatalog function 
(required columns: snp_id, chr, pos)}

\item{vcf_file}{Genotype data}

\item{max_distance}{maximal distance between the qtl variant and the GWAS variant}

\item{min_r2}{Minimal R2 betweeb the QTL vairant and the GWAS variant}
}
\value{
data frame with GWAS and QTL overlaps
}
\description{
Find GWAS variants that are in high LD with QTL variants
}
