% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{findMinimalSnpPvalues}
\alias{findMinimalSnpPvalues}
\title{Find SNP with minimal p-value per gene.}
\usage{
findMinimalSnpPvalues(qtl_df)
}
\arguments{
\item{qtl_df}{Data frame with QTL mapping results 
(required columns: gene_id, p_nominal, n_cis_snps)}
}
\value{
Only gene-SNP pairs with minimal p-values,
 added columns: p_bonferroni, p_fdr.
}
\description{
Group QTL matrix by gene, sort by p_nominal, keep SNP with smalles p-value,
Correct that using bonferroni correction and then apply FDR correction across genes.
}

