% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{gdsToMatrix}
\alias{gdsToMatrix}
\title{Import genotypes in GDS file to a matrix}
\usage{
gdsToMatrix(gds_file)
}
\arguments{
\item{file}{Path to the GDS file.}
}
\value{
List containing snpspos and genotypes matrix.
}
\description{
Open GDS file (created with SNPRelate::snpgdsVCF2GDS) into R and convert it into a matrix of
variant positions and matrix of allele dosage (0,1,2)
}
\author{
Kaur Alasoo
}

