% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_data.R
\name{get.GSE.microarray}
\alias{get.GSE.microarray}
\title{Download and process the microarray data from Gene Expression Omnibus (GEO) by its GSE accession number}
\usage{
get.GSE.microarray(acc_number, take_log = FALSE, ...)
}
\arguments{
\item{acc_number}{the GEO accession number of the microarray data.}

\item{take_log}{whether we want to log2-transform the expression matrix. The default is \code{FALSE}.}

\item{...}{other parameters to be passed to \link[GEOquery:getGEO]{GEOquery::getGEO}.}
}
\value{
a list containing two fields: \code{data_matrix} gives the genotype expression data in a matrix form;
\code{p_data} gives the phenotype data in a data frame.
}
\description{
This function downloads and processes microarray gene expression data stored in the GSE format from GEO into a count matrix using
its accession number. This function makes use of the GEOquery package built by Sean Davis.
}
\examples{
ls <- get.GSE.microarray('GSE61754')
ls

}
