% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_postProcess.R
\name{getFDR}
\alias{getFDR}
\title{Calculate FDR based on an empirical null distribution}
\usage{
getFDR(p1, p0, alpha = 0.1, z = NULL, subset = NULL)
}
\arguments{
\item{p1}{alt p-value vector}

\item{p0}{null p-value vector}

\item{alpha}{desired FDR level}

\item{z}{?}

\item{subset}{?}
}
\value{
P-value cutoff corresponding to the desired FDR
}
\description{
Author: Natsuhiko Kumasaka
}
