% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_lists.R
\docType{data}
\name{h1_epitope_positions}
\alias{h1_epitope_positions}
\title{H1 influenza epitope positions}
\format{
An object of class \code{list} of length 1.
}
\usage{
h1_epitope_positions
}
\description{
A list of epitope positions in H1 influenza hemagglutinin (HA).
}
\keyword{datasets}
