% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_lists.R
\docType{data}
\name{h3_epitope_positions}
\alias{h3_epitope_positions}
\title{H3 influenza epitope positions}
\format{
An object of class \code{list} of length 7.
}
\usage{
h3_epitope_positions
}
\description{
A list of epitope positions in H3 influenza hemagglutinin (HA).
}
\keyword{datasets}
