% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_classify_peaks.R
\name{identifyAmbiguousMasters}
\alias{identifyAmbiguousMasters}
\title{Identify master peaks that share credible sets with each other}
\usage{
identifyAmbiguousMasters(shared_credible_sets, master_peak_list)
}
\description{
Identify master peaks that share credible sets with each other
}
