% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_classify_peaks.R
\name{importCredibleSets}
\alias{importCredibleSets}
\title{Import credible sets from disk and add overlapping peaks to each QTL}
\usage{
importCredibleSets(credible_set_path, peak_metadata)
}
\description{
Import credible sets from disk and add overlapping peaks to each QTL
}
