% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_gwas.R
\name{importGwasCatalog}
\alias{importGwasCatalog}
\title{Import EBI GWAS catalog into R.}
\usage{
importGwasCatalog(path)
}
\arguments{
\item{path}{Path to the EBI GWAS catalog export file (must contain MAPPED_TRAIT field as well).}
}
\value{
data frame containing the gwas catalog.
}
\description{
Imports tsv version of the GWAS catalog into R. Parses total sample size from the 
INITIAL SAMPLE DESCRIPTION field and also flags each study that contains European samples.
}
\author{
Kaur Alasoo
}

