% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{importRasqualTable}
\alias{importRasqualTable}
\title{Import rasqual output table into R}
\usage{
importRasqualTable(path)
}
\arguments{
\item{path}{Bath to rasqual output file.}
}
\value{
data_frame
}
\description{
Skipped gene-SNP pairs are automatically removed and 
chisq statistic is converted into p-value (p_nominal).
}

