% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{importVariantInformation}
\alias{importVariantInformation}
\title{Import variant information extracted from VCF file into R}
\usage{
importVariantInformation(path)
}
\description{
Import variant information extracted from VCF file into R
}
