% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{loadCounts}
\alias{loadCounts}
\title{Import featureCounts output table into R.}
\usage{
loadCounts(sample_dir, sample_names, counts_suffix = ".counts.txt",
  sub_dir = TRUE)
}
\arguments{
\item{sample_dir}{Path to the directory containing the count files.}

\item{sample_names}{Vector of sample names.}

\item{counts_suffix}{Suffix of the counts file.}

\item{sub_dir}{If TRUE, count files are nested in subfolders named after sample names, otherwise counts files
are directly in sample_dir.}
}
\value{
Counts matrix where the first two columns are gene_id and feature length.
}
\description{
Skips the first comment line.
}
\author{
Kaur Alasoo
}
