% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_data.R
\name{loadNarrowPeaks}
\alias{loadNarrowPeaks}
\title{Import narrowPeak files into a list of GRanges objects.}
\usage{
loadNarrowPeaks(sample_dir, sample_names, peaks_suffix = "_peaks.narrowPeak")
}
\arguments{
\item{sample_dir}{Path to the directory containing the narroPeak files.}

\item{sample_names}{Vector of sample names.}

\item{peaks_suffix}{Suffix of the narrowPeak files.}
}
\value{
List of GRanges objects corresponding to peak calls from each sample.
}
\description{
All narrowPeak files are assumed to be nested in subdirectories named adter sample names.
Based on rtracklayer::import.bed, but adds additional columns present only in narrowPeak files.
}
\author{
Kaur Alasoo
}

