% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot_fun.R
\name{make.heatmap}
\alias{make.heatmap}
\title{Plot a Heatmap for an Expression Value Matrix}
\usage{
make.heatmap(
  sig_mat,
  plot_title = "Signature Matrix Heatmap",
  names = colnames(sig_name)
)
}
\arguments{
\item{sig_mat}{The matrix for which the heatmap is plotted.}

\item{plot_title}{Custom title for thie plot. Default is 'Signature Matrix Heatmap'.}

\item{names}{The names of the columns of the final heatmap. Default is \code{colnames(sig_name)}.}
}
\description{
This function will plot a heatmap for an expression value matrix. It will automatically group genes
that are highly expressed for each column so the diagonal blocks of the heatmap show the highest expression values.
This function will be particularly helpful for constructing a heatmap of a signature matrix or a differential expression map.
}
