% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot_fun.R
\name{make.theta.heat.map}
\alias{make.theta.heat.map}
\title{Function to Produce The Heatmap from a Matrix Using \code{ggplot2}}
\usage{
make.theta.heat.map(Theta_star, title, all_names)
}
\description{
Function to Produce The Heatmap from a Matrix Using \code{ggplot2}
}
