% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtl_makePlots.R
\name{makeMultiplePlots}
\alias{makeMultiplePlots}
\title{Make a list of plotEQTL plots.}
\usage{
makeMultiplePlots(snps_df, expression_matrix, genotype_matrix, sample_metadata,
  gene_metadata)
}
\arguments{
\item{snps_df}{Data frame with at least two columns (gene_id, snp_id) corresponding to snps and genes to be plotted.}

\item{expression_matrix}{Matrix of normalised gene expression values (genes in rows, samples in columns).
Column names correspond to sample_id in sample_metadata.}

\item{genotype_matrix}{Matrix of genotypes (variants in rows, individuals in columns).
Column names correspond to genotype_id in sample_metadata.}

\item{sample_metadata}{data frame linking samples to genotypes (Required columns: sample_id, genotype_id, condition_name)}

\item{gene_metadata}{data frame for linking gene ids to gene names. (Required columns: gene_id, gene_name)}
}
\value{
List of ggplot2 objects.
}
\description{
Plots are faceted by condition_name column in sample_metadata.
}
\author{
Kaur Alasoo
}

