% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_processPeaks.R
\name{makeUnionPeaks}
\alias{makeUnionPeaks}
\title{Construct the union of list of peak calls and add gene_id and type metadata columns for read counting.}
\usage{
makeUnionPeaks(peak_list, seqlevels, id_prefix)
}
\description{
Construct the union of list of peak calls and add gene_id and type metadata columns for read counting.
}
