% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proc_fun.R
\name{man.filter}
\alias{man.filter}
\title{RNA Sequencing Data Manual Filtering}
\usage{
man.filter(raw, min_sum = 10, min_var = 1e-04, verbose = TRUE)
}
\arguments{
\item{raw}{raw sequencing data in the wide format}

\item{min_sum}{minimal sum of gene expression over all samples to be allowed in the data; default value is \code{10}}

\item{min_var}{minimal gene expression variance over all samples to be allowed in the data; default value is \code{1e-4}}

\item{verbose}{logical value to indicate whether detailed information of filtering should be printed; default value is \code{TRUE}}
}
\value{
a filtered raw sequencing data based on the specified values
}
\description{
Manual filtering of unqualified rows from an RNA-seq dataset. The parameters are set
for bulk RNA seq data. Four filtering steps are performed to remove unqualified rows:
\itemize{
\item Remove non-genes (i.e. rows for summary statistics, etc.)
\item Remove genes with low expression across samples (sum of expression < \code{min_sum})
\item Remove genes with low variance across samples (variance < \code{min_var})
\item Remove duplicate rows (i.e. rows with duplicate gene names, etc.)
}
}
\details{
Although software like DESeq2 has steps like independent filtering that allows users
to skip manual pre-filtering, it is still recommended to do so sometimes.
}
\examples{
data(example)
example[1, ] = runif(ncol(example), 0, 0.01) # to create an example gene with low expression
example[2, ] = 15 # an example of gene with no variation
d1 = man.filter(example) # default setting
d2 = man.filter(example, min_sum = 0) # keep the new row 1
d3 = man.filter(example, min_var = 0) # keep the new row 2
d4 = man.filter(example, verbose = FALSE) # same as d1 in the output but no information printed

}
