% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_matrixEQTL.R
\name{matrixeqtlTestInteraction}
\alias{matrixeqtlTestInteraction}
\title{Use MatrixEQTL to test interaction between genotype and two conditions}
\usage{
matrixeqtlTestInteraction(conditions, gene_snp_pairs, expression_list,
  genotype_list, covariate_names = NULL)
}
\arguments{
\item{conditions}{Character vector containing the names of the two conditions}

\item{gene_snp_pairs}{Data frame with two columns: gene_id, snp_id.}

\item{expression_list}{List with expression data.}

\item{genotype_list}{List with genotype data.}

\item{covariates_names}{Names of the columns in the expression_list$sample_metadata table 
that will be used as covariates for the matrixEQTL model.}
}
\value{
Tidy table with interaction testing results
}
\description{
Use MatrixEQTL to test interaction between genotype and two conditions
}

