% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtl_makePlots.R
\name{plotEQTL}
\alias{plotEQTL}
\title{Make a boxplot of expression QTL in multiple conditions.}
\usage{
plotEQTL(selected_gene_id, genotype_id, expression_matrix, genotype_matrix,
  sample_metadata, gene_metadata)
}
\arguments{
\item{selected_gene_id}{ID if the gene of interest.}

\item{expression_matrix}{Matrix of normalised gene expression values (genes in rows, samples in columns).
Column names correspond to sample_id in sample_metadata.}

\item{genotype_matrix}{Matrix of genotypes (variants in rows, individuals in columns).
Column names correspond to genotype_id in sample_metadata.}

\item{sample_metadata}{data frame linking samples to genotypes (Required columns: sample_id, genotype_id, condition_name)}

\item{gene_metadata}{data frame for linking gene ids to gene names. (Required columns: gene_id, gene_name)}

\item{gentype_id}{ID of the variant of interest.}
}
\value{
ggplot2 object
}
\description{
Plots are faceted by condition_name column in sample_metadata.
}
\author{
Kaur Alasoo
}

