% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_fastqtl.R
\name{prepareFastqtlMatrix}
\alias{prepareFastqtlMatrix}
\title{Add gene position to the FastQTL expression matrix}
\usage{
prepareFastqtlMatrix(matrix, genepos)
}
\arguments{
\item{matrix}{expression matrix}

\item{genepos}{data frame form constructFastQTLGenePos}
}
\value{
data_frame suitable for FastQTL
}
\description{
Add gene position to the FastQTL expression matrix
}
\author{
Kaur Alasoo
}
