% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{prepareQTLtoolsMatrix}
\alias{prepareQTLtoolsMatrix}
\title{Add gene position to the QTLtools expression matrix}
\usage{
prepareQTLtoolsMatrix(matrix, genepos)
}
\arguments{
\item{matrix}{expression matrix}

\item{genepos}{data frame form constructQTLtoolsGenePos}
}
\value{
data_frame suitable for QTLtools
}
\description{
Add gene position to the QTLtools expression matrix
}
\author{
Kaur Alasoo
}
