% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_FINEMAP.R
\name{pToZ}
\alias{pToZ}
\title{Convert FastQTL p-values into a data frame of z-scores}
\usage{
pToZ(pvalue_df, n_top_snps = 200)
}
\arguments{
\item{pvalue_df}{data.fram of p-values (columns: p_nominal, beta)}

\item{n_top_snps}{Number of top associated SNPs to be included.}
}
\value{
data frame of z-scores
}
\description{
Convert FastQTL p-values into a data frame of z-scores
}

