% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot_fun.R
\name{pw.scatter}
\alias{pw.scatter}
\title{Plot Pairwise Scatterplot for Samples}
\usage{
pw.scatter(Y, notes = NULL, dir_name = ".", par_param = c(3, 3), ...)
}
\arguments{
\item{Y}{a matrix of expression values with row names corresponding to genes and column names to sample names}

\item{notes}{additional notes for file names, such as generations, platforms, sample information, etc. Default is \code{NULL}}

\item{dir_name}{name of the directory to which the output jpeg files are saved. Default is the working directory (\code{getwd()})}

\item{par_param}{a vector of 2 specifying the number of rows and columns of the output plots. Default is \code{c(3, 3)}}

\item{...}{other parameters for the \link[base:plot]{base::plot} function}
}
\value{
output no direct return but save the pairwise scatterplots in a jpeg file to the named directory
}
\description{
Plot Pairwise Scatterplot for Samples
}
\examples{
data(example)
pw.scatter(Y = example[, 1:5])
pw.scatter(Y = example[, 1:5], notes = 'example')
pw.scatter(Y = example[, 1:5], notes = 'example', par_param = c(2, 3))

}
