\name{plot}
\Rdversion{1.1}
\docType{methods}
\alias{plot}
\alias{plot,ANY}
\alias{plot,QAReadQuality-method}
\alias{plot,QANucleotideByCycle-method}
\alias{plot,QAQualityByCycle-method}
\alias{plot,QANucleotideUse-method}
\alias{plot,QAQualityUse-method}
\alias{plot,QASequenceUse-method}
\alias{plotDensity}
\alias{plotDensity,QAQualityUse-method}
\alias{table}
\alias{table,ANY}
\alias{table,QANucleotideUse-method}
\alias{table,QAFrequentSequence-method}
\alias{table,QAAdapterContamination-method}

\title{Plotting functions for quality objects}
\description{
  These functions allow to plot or retrieve summary statistics for
  quality assessment objects generated by the \code{qa2} function of the
  \code{ShortRead} package.
}
\usage{

\S4method{plot}{QAReadQuality}(x, y, ... )

\S4method{plot}{QANucleotideByCycle}(x, y, dnaCol=c("#1F78B4",
"#33A02C", "#B2DF8A", "#A6CEE3", "#E31A1C"), ... )

\S4method{plot}{QAQualityByCycle}(x, y, type="levelplot", ... )

\S4method{plot}{QANucleotideUse}(x, y, dnaCol=c("#1F78B4", "#33A02C",
"#B2DF8A", "#A6CEE3", "#E31A1C"), ...)

\S4method{plot}{QAQualityUse}( x, y, ... )

\S4method{plot}{QASequenceUse}( x, y, ... )

\S4method{plotDensity}{QAQualityUse}( x, xlab="Used base-call quality",
ylab="density", col, pal=brewer.pal( 8, "Set1" ), lty=1, ... )

\S4method{table}{QANucleotideUse}( x, ... )

\S4method{table}{QAFrequentSequence}( x, ... )

\S4method{table}{QAAdapterContamination}( x, ... )

}
\arguments{

  \item{...}{
    Additional arguments.
  }

  \item{col}{
    Allows to specify the colors for the individual sequencing runs manually.
  }

  \item{dnaCol}{
    The colors that for the individual nucleotides (A, C, G, T and N).
  }

  \item{lty}{
    For \code{plotDensity}: the line type.
  }

  \item{type}{
    For \code{QAQualityByCycle}: allowed values: \code{"levelplot"} and
    \code{"boxplot"}. If the base call quality information
    per cycle should be visualized as a levelplot or boxplot.
  }

  \item{pal}{
    Vector of colors that can be used to define the colors for the
    individual sequencing runs. Colors might be recycled.
  }

  \item{x}{
    The QA object (see function \code{\link[ShortRead]{qa2}} in the \code{ShortRead} package).
  }

  \item{xlab}{
    For \code{plotDensity}: the x-axis label.
  }

  \item{y}{
    Not used by any of the functions.
  }

  \item{ylab}{
    For \code{plotDensity}: the y-axis label.
  }
}
\section{Methods and Functions}{
  \describe{
    \item{plot}{
      Creates a plot summarizing the quality information from the
      corresponding QA object (see \code{\link[ShortRead]{qa2}} for a
      full list of QA objects).
    }

    \code{plotDensity}{
      Draws a density plot representing the signal distribution.
    }

    \item{table}{
      returns summary statistics for the corresponding QA object:
      \itemize{
	\item{For \code{QAFrequentSequence}:}{
	  A \code{data.frame} with columns \code{Id}, \code{Records},
	  \code{Seq_count} and \code{Sequence} for the sample ID, the number
	  of records (reads) for that sample, the number of occurences of
	  the sequence and the sequence.
	}

	\item{For \code{QANucleotideUse}:}{
	  A \code{data.frame} with the total count of nucleotides as
	  well as the percentage of nucleotide use in all sequences of
	  the sample.
	}

      }
    }
  }
}
\author{
  Johannes Rainer
}
\seealso{
  \code{\link[ShortRead]{qa2}}
}
\examples{

#############
## Examples:
##
## Perform a quality assessment on a fastq file:
dirpath <- system.file( package="SeqUtils", "extdata" )
fastqs <- dir( dirpath, "example.fastq.gz", full=TRUE )

## collating the quality controls:
coll <- QACollate(
    QAFastqSource( con=fastqs, n=10e+06 ),
    QAReadQuality(),
    QAAdapterContamination(),
    QANucleotideUse(),
    QAQualityUse(),
    QASequenceUse(),
    QAFrequentSequence(n=10),
    QANucleotideByCycle(),
    QAQualityByCycle()
    )
## perform the quality control
qualy <- qa2( coll, verbose = TRUE )


############
## distribution of average base call quality per read
plot( qualy$QAReadQuality )


############
## nucleotide quality by cycle:
plot( qualy$QANucleotideByCycle )


###########
## base call quality by cycle
plot( qualy$QAQualityByCycle )
plot( qualy$QAQualityByCycle, type="boxplot" )


###########
## nucleotide use
plot( qualy$QANucleotideUse )

## in addition return the table with the actual values (and percentages)
table( qualy$QANucleotideUse )


##########
## base call quality use
plot( qualy$QAQualityUse )

## plot the base call quality as density distributions
plotDensity( qualy$QAQualityUse )


##########
## sequence use
plot( qualy$QASequenceUse )


#########
## frequent sequence
table( qualy$QAFrequentSequence )


#########
## adapter contamination
table( qualy$QAAdapterContamination )


}
\keyword{classes}

