% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment.R
\name{quantifyMotifDisruption}
\alias{quantifyMotifDisruption}
\title{Estimate how much PWM binding score is altered by the SNP}
\usage{
quantifyMotifDisruption(pwm, peak_id, snp_id, peak_metadata, peak_sequences,
  snp_metadata, window_size = 25)
}
\arguments{
\item{pwm}{Position weight matrix in TFBSTools format}

\item{peak_id}{ID of the ATAC peak}

\item{snp_id}{ID of the variant}

\item{peak_metadata}{Peak metadata (required columns: gene_id, start)}

\item{peak_sequences}{DNAStringSet object containing the peak sequences.}

\item{snp_metadata}{SNP metadata (required columns: snp_id, chr, pos, ref, alt)}

\item{window_size}{Window size around the SNP}
}
\value{
Data frame of binding socres for the reference and alternate versions of the sequence.
}
\description{
Estimate how much PWM binding score is altered by the SNP
}
