% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment.R
\name{quantifyMultipleMotifs}
\alias{quantifyMultipleMotifs}
\title{Applies quantifyMotifDisruption to a list of motifs}
\usage{
quantifyMultipleMotifs(peak_id, snp_id, pwm_list, peak_metadata, peak_sequences,
  snp_metadata, window_size = 25, max_score_thresh = 0.7,
  rel_diff_thresh = 0)
}
\arguments{
\item{max_score_thresh}{Maximal relative score across variants must be greater than this (default = 0.8)}

\item{rel_diff_thresh}{Difference in relative binding score must be greater than this (default = 0)}
}
\description{
Converts the output into a data frame and performs some basic filtering
}

