% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual_effects.R
\name{rankAtacSummaries}
\alias{rankAtacSummaries}
\title{Add gene_id and gene_name to atac hits and rank them by "rank_by" column}
\usage{
rankAtacSummaries(atac_beta_summaries, joint_effects, rank_by)
}
\description{
Add gene_id and gene_name to atac hits and rank them by "rank_by" column
}
