% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{rasqualCalculateSampleOffsets}
\alias{rasqualCalculateSampleOffsets}
\title{Calculate sample-specic offsets for RASQUAL}
\usage{
rasqualCalculateSampleOffsets(counts, gene_metadata, method = "library_size",
  gc_correct = TRUE)
}
\arguments{
\item{counts}{Matrix of read counts (genes in rows, samples in columns)}

\item{gene_metadata}{Matrix of gene metadata (required columns: gene_id, precentage_gc_content). 
Used for GC content bias correction.}

\item{gc_correct}{If true then correct for GC content bias in addition to library size.}
}
\value{
Matrix of gene-specifc offsets.
}
\description{
Calculate offset values for each gene in each condition that
 correct for library size and GC content bias.
}

