% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{rasqualConstructGeneBatches}
\alias{rasqualConstructGeneBatches}
\title{Split gene ids into batches for runRasqual.py script}
\usage{
rasqualConstructGeneBatches(gene_metadata, batch_size)
}
\arguments{
\item{gene_metadata}{Data frame with gene metadata (gene_id column required)}

\item{batch_size}{Number of genes in a batch}
}
\value{
Data frame of gene batches (columns: batch_id, gene_ids)
}
\description{
Split gene ids into batches for runRasqual.py script
}

