% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{rasqualGcCorrection}
\alias{rasqualGcCorrection}
\title{Estimate the effect of GC-bias for each feature in each sample.}
\usage{
rasqualGcCorrection(Y, gene_metadata, PLOT = F)
}
\arguments{
\item{Y}{Matrix of read counts}

\item{gene_metadata}{Data frame with gene metadata.
(required columnss: gene_id, percentage_gc_content)}

\item{PLOT}{}
}
\value{
Matrix of GC-bias offsets for each gene in each condition.
}
\description{
This function does not correct for differences in library size between samples.
}
\author{
Natsuhiko Kumasaka
}

