% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{rasqualOptimisedGeneBatches}
\alias{rasqualOptimisedGeneBatches}
\title{Split genes into different batch sizes based on how many cis and feature SNPs they have.}
\usage{
rasqualOptimisedGeneBatches(gene_metadata, batch_sizes = c(20, 8, 3, 1),
  batch_prefix = "batch")
}
\arguments{
\item{gene_metadata}{Data frame with gene meta data, required columns: gene_id, feature_snp_count, cis_snp_count.}

\item{batch_sizes}{Vector of length 4. Contains the number of gene in a batch for each of the four groups.}

\item{batch_prefix}{Prefix of the batch id, default "batch".}
}
\value{
Data frame of genes split into batches.
}
\description{
This script calculates the number of tests required for each gene (feature_snp_count * cis_snp_count)
and splits genes into four groups based on this value (<15,000; 15,000 to 50,000; 50,000 to 100,000 and > 100,000).
}

