% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_sequences.R
\docType{data}
\name{references}
\alias{references}
\title{Reference GenBank files}
\format{
An object of class \code{list} of length 2.
}
\usage{
references
}
\description{
A nested list of file paths to reference GenBank files bundled with the
package. Access files like \code{references$H3N2$HA}.
}
\keyword{datasets}
