% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtl_processData.R
\name{runMatrixEQTL}
\alias{runMatrixEQTL}
\title{A wrapper around MatrixeQTL}
\usage{
runMatrixEQTL(exp_data, geno_data, snpspos, genepos, covariates = NULL)
}
\arguments{
\item{exp_data}{Matrix of gene expression data (genes in rows, samples in columns).}

\item{geno_data}{Matrix of genotype data (SNPs in rows, samples in columns).}

\item{snpspos}{Matrix of SNP coordinates (columns: snpid,chr,pos).}

\item{genepos}{Matrix of gene coordinates (columns: geneid,chr,left,right).}

\item{covariates}{Matrix of covariates (samples in columns).}
}
\value{
MatrixeQTL result object.
}
\description{
Prepares SlicedData objects of expression data, genotypes and covariates
and then runs MatrixeQTL on the data.
}
\author{
Kaur Alasoo
}

