\name{sampleBam}
\Rdversion{1.1}
\docType{methods}
\alias{sampleBam}
\alias{sampleBam,character-method}
\alias{sampleBam,BamFile-method}


\title{Extracting random samples from a BAM file}
\description{
  This function extracts random samples from a BAM file. It uses the
  \code{samtools view} function, thus \code{samtools} have to be
  installed. Samtool's \code{view -s} option selects reads according to
  scores based on a hash calculated from their read names. Hence whether
  reads are sampled depends on the name of the read which means that
  read pairs are sampled together (if they have the same name).
}
\usage{

\S4method{sampleBam}{character}( x, seed=sample( 1:1e+6, 1 ),
 fraction=0.001, outfile=tempfile(), ... )

\S4method{sampleBam}{BamFile}( x, seed=sample( 1:1e+6, 1 ),
fraction=0.001, outfile=tempfile(), ... )

}
\arguments{

  \item{...}{
    Additional arguments to be passed to the
    \code{\link[Rsamtools]{scanBam}} function.
  }

  \item{fraction}{
    The proportion of reads to be sampled from the BAM file.
  }

  \item{outfile}{
    The file name of the intermediate BAM file that is generated by the
    \code{samtools view -s} call. This file will be deleted once it is read.
  }

  \item{seed}{
    The seed for the random number generator.
  }

  \item{x}{
    The BAM file name or a \code{BamFile} object.
  }

}
\section{Methods and Functions}{
  \describe{
    \item{sampleBam}{
      the function samples the specified fraction of reads from the Bam
      file. The returned object is a list of lists. For more details
      please see the help of the \code{\link[Rsamtools]{scanBam}} function.
    }
  }
}
\author{
  Johannes Rainer
}
\seealso{
  \code{\link[Rsamtools]{scanBam}}
}
\examples{

#############
## Examples:
##
## Perform a quality assessment on a fastq file:
dirpath <- system.file( package="SeqUtils", "extdata" )
bamfile <- dir( dirpath, "example.bam", full=TRUE )

BamSample <- sampleBam( bamfile, fraction=0.01 )
BamSample

}
\keyword{classes}

