% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_eqtlbma.R
\name{saveEqtlbmaData}
\alias{saveEqtlbmaData}
\title{Save eqtl_data_list into files suitable for eqtlbma}
\usage{
saveEqtlbmaData(eqtl_data_list, output_dir, project_root)
}
\arguments{
\item{eqtl_data_list}{List of all data required for eQTL mapping with eqtlbma.}

\item{output_dir}{relative path to the output dir}

\item{project_root}{Absolute path to the prohect root directory.}
}
\value{
None
}
\description{
Helper function to export processed eQTL data into multiple text files required by eqtlbma.
}
\author{
Kaur Alasoo
}

