% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtl_eqtlbma.R
\name{saveEqtlbmaMatrices}
\alias{saveEqtlbmaMatrices}
\title{Save a list of matrices into a suitable format for eqtlbma}
\usage{
saveEqtlbmaMatrices(data_list, output_dir, file_suffix = "exression",
  project_root)
}
\arguments{
\item{data_list}{list of matrices}

\item{output_dir}{relative path to the output dir}

\item{file_suffix}{suffix added to each file after their name in the list.}

\item{project_root}{Absolute path to the prohect root directory.}
}
\value{
None
}
\description{
Works with expression, genotypes and covariates. Also creates a text file with absolute paths to each individual file.
}
\author{
Kaur Alasoo
}

