% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_PEER.R
\name{savePEERData}
\alias{savePEERData}
\title{Save list of expression matrices in a format suitable for runPEER.py}
\usage{
savePEERData(expression_list, outdir, file_suffix = "exprs")
}
\arguments{
\item{expression_list}{list of expression matrices}

\item{outdir}{output directory}

\item{file_suffix}{suffix of the output file}
}
\description{
Save list of expression matrices in a format suitable for runPEER.py
}

