% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_makePlots.R
\name{savePlots}
\alias{savePlots}
\title{Save a list of ggplot2 plots into a folder}
\usage{
savePlots(plot_list, path, width, height)
}
\arguments{
\item{plot_list}{List of ggplot2 objects.}

\item{path}{Path to the output directory, created automatically if does not exist.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}
}
\value{
None
}
\description{
File name is extracted automatically from the plot title.
}
\author{
Kaur Alasoo
}
