% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{saveRasqualMatrices}
\alias{saveRasqualMatrices}
\title{Save a list of matrices into a suitable format for RASQUAL}
\usage{
saveRasqualMatrices(data_list, output_dir, file_suffix = "expression")
}
\arguments{
\item{data_list}{list of matrices}

\item{output_dir}{relative path to the output dir}

\item{file_suffix}{suffix added to each file after their name in the list.}
}
\value{
None
}
\description{
Works with expression and covariates matrices.
}
\author{
Kaur Alasoo
}

