% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqUtils-package.R
\docType{package}
\name{seqUtils-package}
\alias{seqUtils}
\alias{seqUtils-package}
\title{seqUtils: Sequence tools}
\description{
Tools for working with sequence data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamT123/seqUtils}
  \item Report bugs at \url{https://github.com/SamT123/seqUtils/issues}
}

}
\author{
\strong{Maintainer}: Sam Turner \email{sat65@cam.ac.uk}

}
\keyword{internal}
