% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{splitIntoBatches}
\alias{splitIntoBatches}
\title{Split vector of n elements into batches}
\usage{
splitIntoBatches(n, batch_size)
}
\arguments{
\item{n}{Length of the vector}

\item{batch_size}{Size of each batch}
}
\value{
vector of integers where each integer occurs at most batch_size times.
}
\description{
Given a total number of elements n and batch_size, contruct a vector of length n where
}
\examples{
splitIntoBatches(11,3)
}

