% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_enrichment.R
\name{subsampleFisherTest}
\alias{subsampleFisherTest}
\title{Perform Fisher's Exact test when sample is a subsample of a larger population}
\usage{
subsampleFisherTest(sample_hits, sample_size, total_hits, total_size)
}
\arguments{
\item{sample_hits}{number of hits in the sample}

\item{sample_size}{sample size}

\item{total_hits}{number of hits in the total populaton}

\item{total_size}{population size}
}
\value{
data frame with p-value, OR and confidence intervals
}
\description{
Return a tidy data frame with results (p-value, OR, confidence intervals and their log2 coversions)
}
