% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{tabixFetchGenes}
\alias{tabixFetchGenes}
\title{Fetch particular genes from tabix indexed Rasqual output file.}
\usage{
tabixFetchGenes(gene_ranges, tabix_file)
}
\arguments{
\item{gene_ranges}{GRanges object with coordinates of the cis regions around genes.}

\item{tabix_file}{Tabix-indexed Rasqual output file.}
}
\value{
List of data frames containing Rasqual results for each gene.
}
\description{
Fetch particular genes from tabix indexed Rasqual output file.
}

