% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_coloc.R
\name{testColoc}
\alias{testColoc}
\title{Wrapper around coloc.abf to directly work with tidy data frames}
\usage{
testColoc(df1, df2, n1, n2, p1 = 1e-04, p2 = 1e-04, p12 = 1e-05)
}
\arguments{
\item{df1}{Summary statistics from the first dataset
(required columns: snp_id, p_nominal, beta, MAF).}

\item{n1}{Sample size of the first dataset.}

\item{n2}{Sample size of the second dataset.}

\item{p1}{prior probability a SNP is associated with trait 1, default 1e-4}

\item{p2}{prior probability a SNP is associated with trait 2, default 1e-4}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5}

\item{df1}{Summary statistics from the second dataset
(required columns: snp_id, p_nominal, beta, MAF).}
}
\value{
List of colocalisation results (Same as coloc.abf).
}
\description{
Group QTL database by gene, sort by p_nominal, keep SNP with smallest p-value,
correct that using bonferroni correction and then apply FDR correction across genes.
}
