% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_test_interactions.R
\name{testThreewayInteraction}
\alias{testThreewayInteraction}
\title{Test for threeway interactions between two features, genotype and environment.}
\usage{
testThreewayInteraction(feature_pairs, trait_matrix, sample_metadata, vcf_file,
  model0, model1, p_only = TRUE)
}
\arguments{
\item{feature_pairs}{data frame with master_id, dependent_id and snp_id}

\item{trait_matrix}{expression or chromatin accessibility matrix}

\item{sample_metadata}{data frame with sample metadata and covariates}

\item{vcf_file}{VCF file from gdsToMatrix() function}

\item{model0}{Null model}

\item{model1}{Model of interest that is compared to the null model}

\item{p_only}{Specifies return type}
}
\value{
If TRUE, then only p-value is returned, otherwise returns a list of model data and fits.
}
\description{
The two features (master and dependent) are assumed to be regulated by the same 
causal genetic variant.
}
