% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{tibbleToNamedMatrix}
\alias{tibbleToNamedMatrix}
\title{Convert a data frame with an id column into a matrix with row names}
\usage{
tibbleToNamedMatrix(tibble, row_names = "transcript_id")
}
\description{
Convert a data frame with an id column into a matrix with row names
}

