% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_ASE.R
\name{tidyGenotypeMatrix}
\alias{tidyGenotypeMatrix}
\title{Convert genotype matrix from gdsToMatrix into a tidy data frame}
\usage{
tidyGenotypeMatrix(genotypes, selected_snp_id = "snp_id",
  value_id = "snp_value")
}
\arguments{
\item{genotypes}{Matrix with variants in rows and individuals in columms.}

\item{selected_snp_id}{Name for the snp_id column (default: snp_id).}

\item{value_id}{Name for the allele dosage column (default: snp_value).}
}
\value{
Tidy data frame with one genotype per row.
}
\description{
Convert genotype matrix from gdsToMatrix into a tidy data frame
}

