% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/counts_processCounts.R
\name{tidyVector}
\alias{tidyVector}
\title{Convert named vector into a tidy data_frame.}
\usage{
tidyVector(named_vector)
}
\arguments{
\item{named_vector}{Named vector.}
}
\value{
data_frame with two columns: value, sample_id.
}
\description{
Convert named vector into a tidy data_frame.
}
\author{
Kaur Alasoo
}

