% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCompFunctions.R
\name{varianceExplained}
\alias{varianceExplained}
\title{Calculate the proportion of variance explaned by different factors in a lme4 model}
\usage{
varianceExplained(lmer_model)
}
\description{
Calculate the proportion of variance explaned by different factors in a lme4 model
}

