% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_makePlots.R
\name{wiggleplotrConstructMetadata}
\alias{wiggleplotrConstructMetadata}
\title{Construct metadata data frame for wiggleplotr}
\usage{
wiggleplotrConstructMetadata(count_matrix, sample_metadata, bigWig_dir,
  condition_name_levels = c("naive", "IFNg", "SL1344", "IFNg_SL1344"))
}
\arguments{
\item{count_matrix}{Feature count matrix used to for calculating library size}

\item{sample_metadata}{Sample metadata matrix (required columns: sample_id, genotype_id, condition_name)}

\item{bigWig_dir}{Path to the directory with bigWig files}

\item{condition_name_levels}{Levels of the condition name factor}
}
\value{
Wiggleplotr metadata data frame
}
\description{
Construct metadata data frame for wiggleplotr
}

