% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_FINEMAP.R
\name{zToLD}
\alias{zToLD}
\title{Calculate LD matrix based on a data frame of z-scores and genotypes}
\usage{
zToLD(z_df, genotypes)
}
\arguments{
\item{z_df}{data frame of z-scores}

\item{genotypes}{matrix of genotypes}
}
\value{
matrix of pearson correlaton between genotypes
}
\description{
Calculate LD matrix based on a data frame of z-scores and genotypes
}
