% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substitutions.R
\name{get_substitutions}
\alias{get_substitutions}
\title{Get substitutions which differ between two sequences}
\usage{
get_substitutions(
  sequence_1,
  sequence_2,
  position_map = 1:nchar(sequence_1),
  exclude = c(),
  simplify = T
)
}
\arguments{
\item{sequence_1}{"from" sequnce}

\item{sequence_2}{"to" sequence(s); can be multiple sequences}

\item{position_map}{"rename" positions, if you want n'th position of sequences to be referred to with a number other than "n". e.g. if sequence_1 & sequence_2 are just the Koel-7 positions, use \code{position_map = c(145, 155, 156, 158, 159, 189, 193)}}

\item{exclude}{characters to exclude from substitution list (perhaps "X" for aas, "N" for nts)}

\item{simplify}{should output of length 1 (i.e. \code{length(sequence_2)} = 1) be unlisted?}
}
\value{
list of character vectors
}
\description{
Get substitutions which differ between two sequences
}
