% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_fastqtl.R
\name{importFastQTLTable}
\alias{importFastQTLTable}
\title{Import fastQTL output table into R.}
\usage{
importFastQTLTable(file_path)
}
\arguments{
\item{file_path}{Path to the fastQTL output file}
}
\value{
data_frame containing gene_ids, snp ids and p-values.
}
\description{
Detect if the table is from nominal run or permutation run and add proper column names.
}
\author{
Kaur Alasoo
}
