% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{loadBroadPeaks}
\alias{loadBroadPeaks}
\title{Import broadPeak files into a list of GRanges objects.}
\usage{
loadBroadPeaks(sample_dir, sample_names,
  peaks_suffix = "_peaks.broadPeak", sub_dir = TRUE)
}
\arguments{
\item{sample_dir}{Path to the directory containing the broadPeak files.}

\item{sample_names}{Vector of sample names.}

\item{peaks_suffix}{Suffix of the broadPeak files.}

\item{sub_dir}{If TRUE, count files are nested in subfolders named after sample names, otherwise counts files
are directly in sample_dir.}
}
\value{
List of GRanges objects corresponding to peak calls from each sample.
}
\description{
All broadPeak files are assumed to be nested in subdirectories named adter sample names.
Based on rtracklayer::import.bed, but adds additional columns present only in broadPeak files.
}
\author{
Kaur Alasoo
}
