% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafft_align.R
\name{mafft_align}
\alias{mafft_align}
\title{Align sequences using mafft}
\usage{
mafft_align(unaligned_sequences, reference_sequence)
}
\arguments{
\item{unaligned_sequences}{Character vector of unaligned sequences}

\item{reference_sequence}{Single sequence to align \code{unaligned_sequences} to}
}
\value{
Character vector of aligned sequences, with names preserved
}
\description{
Align sequences using mafft
}
