% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot_fun.R
\name{pw.cor.heatmap}
\alias{pw.cor.heatmap}
\title{Plot a Heatmap for Pairwise Expression Correlation}
\usage{
pw.cor.heatmap(Y, coldata)
}
\description{
Plot a Heatmap for Pairwise Expression Correlation
}
