% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_rasqual.R
\name{rasqualGcCorrection}
\alias{rasqualGcCorrection}
\title{Correct rasqual size factors matrix for gc content}
\usage{
rasqualGcCorrection(Y, gcvec, PLOT = F)
}
\arguments{
\item{Y}{Matrix of read counts}

\item{gcvec}{Vector of GC content percentages}

\item{PLOT}{}
}
\value{
Matrix of GC-corrected library sizes.
}
\description{
Correct rasqual size factors matrix for gc content
}
\author{
Natsuhiko Kumasaka
}

