% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_matrixEQTL.R
\name{runMatrixEQTL}
\alias{runMatrixEQTL}
\title{A wrapper around MatrixeQTL}
\usage{
runMatrixEQTL(exp_data, geno_data, snpspos, genepos, covariates = NULL,
  cisDist = 5e+05, pvOutputThreshold = 0.01, permute = FALSE,
  model = modelLINEAR)
}
\arguments{
\item{exp_data}{Matrix of gene expression data (genes in rows, samples in columns).}

\item{geno_data}{Matrix of genotype data (SNPs in rows, samples in columns).}

\item{snpspos}{Matrix of SNP coordinates (columns: snpid,chr,pos).}

\item{genepos}{Matrix of gene coordinates (columns: geneid,chr,left,right).}

\item{covariates}{Matrix of covariates (samples in columns).}

\item{cisDist}{cis distance from the gene.}

\item{pvOutputThreshold}{Maximum p-value to report. Smaller values make the code quicker.}

\item{permute}{Permute genotype labels before qtl mapping.}

\item{model}{Specifies which MatrixEQTL model to use. Options: modelLINEAR, modelLINEAR_CROSS and modelANOVA.}
}
\value{
MatrixeQTL result object.
}
\description{
Prepares SlicedData objects of expression data, genotypes and covariates 
and then runs MatrixeQTL on the data.
}
\author{
Kaur Alasoo
}
