% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqUtils-package.R
\docType{package}
\name{seqUtils-package}
\alias{seqUtils}
\alias{seqUtils-package}
\title{seqUtils: Utilities for Biological Sequence Analysis}
\description{
Provides efficient tools for common sequence analysis tasks including FASTA file I/O, sequence translation, alignment using MAFFT, substitution detection, and consensus calculation. Designed for working with DNA, RNA, and protein sequences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamT123/seqUtils}
  \item Report bugs at \url{https://github.com/SamT123/seqUtils/issues}
}

}
\author{
\strong{Maintainer}: Sam Turner \email{sat65@cam.ac.uk}

}
\keyword{internal}
