% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqUtils-package.R
\docType{package}
\name{seqUtils-package}
\alias{seqUtils}
\alias{seqUtils-package}
\title{seqUtils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamT123/seqUtils}
  \item Report bugs at \url{https://github.com/SamT123/seqUtils/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
