% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_test_interactions.R
\name{testInteraction}
\alias{testInteraction}
\title{Test for interaction between genotype and condition using ANOVA model}
\usage{
testInteraction(gene_id, snp_id, trait_matrix, sample_metadata, vcf_file,
  qtl_formula, interaction_formula, return_value = "ponly")
}
\arguments{
\item{gene_id}{Tested gene id}

\item{snp_id}{Tested SNP id}

\item{trait_matrix}{expression or chromatin accessibility matrix}

\item{sample_metadata}{data frame with sample metadata and covariates}

\item{vcf_file}{VCF file from gdsToMatrix() function}

\item{qtl_formula}{Formula for the model with just genotype and condition terms}

\item{interaction_formula}{Formula for the model with interaction term betweeb genotype and condition}
}
\value{
Either a pvalue (if return_value == "ponly") or the full linear model object.
}
\description{
Test for interaction between genotype and condition using ANOVA model
}
