% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{tidyVector}
\alias{tidyVector}
\title{Convert named vector into a tidy data_frame.}
\usage{
tidyVector(named_vector, value_id = "value", sample_id = "sample_id")
}
\arguments{
\item{named_vector}{Named vector.}
}
\value{
data_frame with two columns: value, sample_id.
}
\description{
Convert named vector into a tidy data_frame.
}
\author{
Kaur Alasoo
}

