% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Wrapper for Biostrings::translate so a character vector can be provided}
\usage{
translate(sequences)
}
\arguments{
\item{sequences}{character vector of DNA/RNA sequences}
}
\value{
character vector of AA sequences, with names preserved
}
\description{
Wrapper for Biostrings::translate so a character vector can be provided
}
