% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Wrapper for Biostrings::translate with deletion handling}
\usage{
translate(sequences, reference_aas = NULL)
}
\arguments{
\item{sequences}{Character vector of DNA/RNA sequences. May contain gap characters ("-").}

\item{reference_aas}{(optional) Reference amino acid sequence. If provided, translated
sequences will be aligned to this reference using MAFFT. This is essential when
input sequences contain deletions and you want to preserve alignment in the output.}
}
\value{
Character vector of amino acid sequences, with names preserved. If \code{reference_aas}
is provided, sequences will be aligned. Otherwise, deletions are simply removed and
sequences may have different lengths.
}
\description{
Translates DNA/RNA sequences to amino acids, with special handling for deletions
(gap characters "-"). The function processes deletions as follows:
}
\details{
\enumerate{
\item Leading deletions are padded to codon boundaries (multiples of 3)
\item All deletions are removed before translation
\item If \code{reference_aas} is provided, translated sequences are realigned to the reference
\item If deletions exist without a reference, output will be unaligned (warning issued)
}
}
\examples{
# Simple translation without deletions
seqs <- c("ATGGCC", "ATGTTT")
translate(seqs)

\dontrun{
# With deletions and reference alignment
seqs <- c("ATGGCC", "ATG---GCC")
ref_aa <- translate(seqs[1])
translate(seqs, reference_aas = ref_aa)
}

}
