% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Wrapper for Biostrings::translate}
\usage{
translate(sequences, reference_aas = NULL)
}
\arguments{
\item{sequences}{Character vector of DNA/RNA sequences}

\item{reference_aas}{(optional) Sequences are aligned to \code{reference} if provided. Useful if, for example, there are deletions in sequences}
}
\value{
character vector of AA sequences, with names preserved
}
\description{
Accepts character vector, and handles deletions
}
